requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.9.1
billiard==4.2.4
blinker==1.9.0
botocore==1.42.33
celery==5.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==5.1
django-celery-beat==2.8.1
django-crontab==0.7.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
drf-yasg==1.21.7
Flask==2.0.0
frozenlist==1.8.0
gql==4.0.0
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.7
graphql-relay==3.2.0
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==3.2.0
kombu==5.6.2
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
Pillow>=10.0.0
promise==2.3
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.9
pyrsistent==0.20.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==5.3.1
redis==7.1.0
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==80.10.2
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.5
yarl==1.22.0
